plugins: rename base file
