+++ core/code/plugin/plugin_base.ts
